Fix propagation of "get-child-position" event
OpacityPreviewToolbox should define sizes and positions for its members
only: opacityPreviewToolbox.widget and eventbox.widget.

This was not the case previously as it could also define allocations for
FloatingToolbox, through the OpacityPreviewToolbox::getOverlayPosition
method.

This commit let the event "get-child-position" propagate if the widget
being positionned in the overlay with OpacityPreviewToolbox::getOverlayPosition
is not one of its members.