Fix segfault caused in updateEventBoxAllocation
The eventbox must be placed over the selected color item.

Hence, to calculate the new allocation, it is necessary to update the
selected color item beforehand.

This was manually done in certain places before calling the method
OpacityPreviewToolbox::updateEventBoxAllocation, but not in others, which led
to occasional segmentation faults by using an old selected color item pointer
that was already deleted."

By including the update of the selected color item in the
updateEventBoxAllocation method, it ensures that the new allocation is
calculated using a fresh pointer to a color item, preventing the use of a
deleted one.