Support duplicated colors in ColorToolItems
Previously, when a color was duplicated in multiple ColorToolItems
(e.g.,when using the floating toolbox or customized toolbars),
the opacity toolbox could appear at a location that did not correspond to
the ColorToolItem under the mouse cursor, but instead matched another
ColorToolItem with the same color.

This commit addresses and resolves this issue.