Resolve Warning in isPointerOverWidget
In OpacityPreviewToolbox::isPointerOverWidget(), there is a potential
issue where the function may be invoked when there is no window associated
with the provided widget parameter.

In such cases, invoking gdk_window_get_origin to determine the window's
coordinates is inappropriate and can lead to issues.

This commit addresses this concern and rectifies the warning.