Turn off eventboxes in floating toolbox if hidden
Invisible eventBoxes that are overlaid on ToolColorItems in the floating
toolbox will no longer emit signals when the floating toolbox is hidden.

Previously, these eventBoxes caused the opacity toolbox to appear when the
mouse entered it, even when the floating toolbox was hidden, thus
preventing the user from interacting normally with the page.

This commit rectifies this behavior, ensuring a seamless experience for
users.