Improve FloatingToolbox visibility
When the mouse enters an eventbox overlaid on a ColorToolItem in the
floatingbox, the opacity toolbox appears, allowing the user to set the
opacity for the current tool.

Previously, leaving the opacity toolbox in this situation did not hide the
floating toolbox as a user would expect, forcing the user to enter it and
exit it to make it disappear.

This commit fixes that issue.